requirements/base.txt
Django==4.2.11
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
djangorestframework-simplejwt==5.3.1
